code/src/backend/requirements.txt
flask-apispec==0.11.0
Flask-RESTful==0.3.9
python-dotenv==0.20.0
gunicorn==20.1.0
flask-mongoengine==1.0.0
mongoengine==0.24.1

